Modified create video method to use aws backend code
Modified the createVid method used to upload a video to
the data lake to perform a get request to recieve an expiring link
to upload to s3 bucket. This link will allow the front end to directly
upload the user's video to the datalake.

Related to: #49
Fixes: #44
Depends on: https://github.com/orate-io/orate-io-backend/pull/32

Signed-off-by: Oussama Saoudi <oussama.sa216@gmail.com>